fix stupid bugs

+++ gemm_hopper_emu_2.h
@@ -8,4 +8,4 @@ constexpr size_t VEC_K = 16;
 #ifdef __cplusplus
 extern "C"
 #endif
-float MulVecVecHopperEmu2(float c, const uint32_t vec_a[VEC_K], const uint32_t vec_b[VEC_K]);
+float MulVecVecHopperEmu2(float c, const uint16_t vec_a[VEC_K], const uint16_t vec_b[VEC_K]);
